requirements.txt
python-dotenv==1.0.1
streamlit==1.40.1
langchain==0.3.11
langchain-community==0.3.11
PyGithub==1.55
pymongo
plotly 
pandas
langchain-openai==0.2.10
langgraph==0.2.53
langchain-core==0.3.29
ipython==8.29.0
openai==1.55.3
langchain-ollama==0.2.2
matplotlib
seaborn
markdown2
paramiko
scikit-learn==1.6.0
torch==2.5.1
InstructorEmbedding==1.0.1
sentence-transformers==2.2.2
huggingface-hub==0.25.0
pypdf==5.1.0
python-docx==1.1.2
chromadb==0.5.20
langchain-chroma==0.1.4